bot/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
tgcrypto = "^1.2.5"
pyrogram = { git = "https://github.com/KurimuzonAkuma/pyrogram.git", tag = "v2.1.31" }
pyrostates = { git = "https://github.com/SajjadAlipour2006/Pyrostates.git" }
uvloop = { version = "^0.21.0", markers = "sys_platform == 'linux'" }
fastapi = "^0.115.5"
uvicorn = "^0.32.1"

[tool.poetry.group.dev.dependencies]
ruff = "0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
